+++ src/main/java/snw/rfm/processor/HunterPauseCardProcessor.java
@@ -21,7 +21,6 @@ import snw.rfm.api.ItemEventListener;
 import snw.rfm.config.ItemConfiguration;
 import snw.rfm.game.GameProcess;
 import snw.rfm.game.TeamHolder;
-import snw.rfm.tasks.HunterReleaseTimer;
 
 public final class HunterPauseCardProcessor implements ItemEventListener, Listener {
     @Override
@@ -30,7 +29,7 @@ public final class HunterPauseCardProcessor implements ItemEventListener, Listen
         if (th.isRunner(player)) { // 排除使用者是猎人从而导致猎人坑队友的情况，哈哈哈哈哈哈哈笑死我了
             int hpctime = ItemConfiguration.getItemTime("hpc");
             GameProcess process = RunForMoney.getInstance().getGameProcess();
-            if (process.getTimers().stream().noneMatch(IT -> IT instanceof HunterReleaseTimer)) {
+            if (process.getHunterReleaseTimer() == null) {
                 process.setHunterNoMoveTime(hpctime);
                 Bukkit.broadcastMessage(ChatColor.GREEN + player.getName() + " 使用了猎人暂停卡。猎人暂停了 " + (hpctime) + " 秒。");
                 return true;