+++ src/main/java/snw/rfm/tasks/HunterReleaseTimer.java
@@ -12,19 +12,24 @@ package snw.rfm.tasks;
 
 import net.md_5.bungee.api.ChatMessageType;
 import net.md_5.bungee.api.chat.TextComponent;
+import org.apache.commons.lang.Validate;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.scheduler.BukkitRunnable;
 import snw.rfm.RunForMoney;
-import snw.rfm.config.GameConfiguration;
+import snw.rfm.game.GameProcess;
 import snw.rfm.game.TeamHolder;
 import snw.rfm.group.GroupHolder;
 
 public final class HunterReleaseTimer extends BaseCountDownTimer {
-    public HunterReleaseTimer(int time) {
+    private final GameProcess process;
+
+    public HunterReleaseTimer(int time, GameProcess process) {
         super(time);
+        Validate.notNull(process);
+        this.process = process;
     }
 
     @Override
@@ -50,11 +55,8 @@ public final class HunterReleaseTimer extends BaseCountDownTimer {
         }
 
         new SendingActionBarMessage(new TextComponent(ChatColor.DARK_RED + "" + ChatColor.BOLD + "猎人已经放出")).start();
-        RunForMoney rfm = RunForMoney.getInstance();
-        CoinTimer ct = new CoinTimer(GameConfiguration.getGameTime() * 60, GameConfiguration.getCoinPerSecond());
-        ct.start(rfm);
-        rfm.getGameProcess().addTimer(ct);
-        rfm.getGameProcess().getTimers().remove(this);
+        process.setHunterReleaseTimer(null);
+        process.getMainTimer().start(RunForMoney.getInstance());
     }
 
     @Override